requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LLM Integration
anthropic>=0.40.0
groq>=0.4.0

# WebSocket
websockets>=12.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Document Export
python-docx>=1.1.0
reportlab>=4.0.0
markdown>=3.5.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
